Move message actions to lib/messages
