+++ apps/web/lib/messages/mark-program-messages-read.ts
@@ -2,7 +2,7 @@
 
 import { prisma } from "@dub/prisma";
 import * as z from "zod/v4";
-import { authPartnerActionClient } from "../safe-action";
+import { authPartnerActionClient } from "../actions/safe-action";
 
 const schema = z.object({
   programSlug: z.string(),
