+++ apps/web/lib/messages/message-partner.ts
@@ -11,9 +11,9 @@ import { prisma } from "@dub/prisma";
 import { APP_DOMAIN_WITH_NGROK } from "@dub/utils";
 import { waitUntil } from "@vercel/functions";
 import * as z from "zod/v4";
-import { MessageSchema, messagePartnerSchema } from "../../messages/schemas";
-import { authActionClient } from "../safe-action";
-import { throwIfNoPermission } from "../throw-if-no-permission";
+import { authActionClient } from "../actions/safe-action";
+import { throwIfNoPermission } from "../actions/throw-if-no-permission";
+import { MessageSchema, messagePartnerSchema } from "./schemas";
 
 const schema = messagePartnerSchema
   .extend({
